requirements.txt
comtypes==1.4.1
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
kiwisolver==1.4.5
matplotlib==3.8.4
MouseInfo==0.1.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
python-dateutil==2.9.0.post0
pytweening==1.2.0
pywin32==306
pywinauto==0.6.8
six==1.16.0
